implement parallel preprocessing
it turned out the implementation of python's multiprocessing.Pool
is exceptionally stupid (for example, it only works with a map function
that's unpractical with anything except an already known list of jobs,
then it blocks indefinitely on Queue.get(), which causes processes to
hang when CTRL-C is issued and the total number of tasks was less than
the size of the jobpool, etc), so i had to implement my own pool code,
which works better and is less than 1/8th the number of lines.

in tests with 8 cpu cores, this causes a speedup of slightly more than 2x.
this is mostly because the .c files i use are already very slim and pre-
processing them is really fast. not optimal yet, but it causes a huge
usability improvement nonetheless. mediumsized programs now build in less
than a second with -j8.