preprocess: on error, rescan if CPPFLAGS pragma was added
preprocessing, even if it fails, can produce results.
a user might want to set some CPPFLAGS with -I include dir directives.
if such a flag is found and it wasn't added already, the file is
scanned again with the new flags.
remove grep pipe, as it causes a possible error code from the pre-
processor to be swallowed.