major revamp of work modes, split into 2 tools
i wasn't really happy with rcb2 lately, because in order to produce an
optimized makefile it needed 3 passes:

1) preprocess to gather documented dependencies
2) build all dependencies with -O0 -g to figure out which
   symbols they provide and remove the unneeded input files.
3) build everything again with the user-provided CFLAGS

not only did that cause much more work than strictly necessary, our
overriding of the CFLAGS for pass 2 also caused the user-provided
CFLAGS to be overridden, which sometimes caused changed build config
and therefore a build error.

i now splitted rcb2 responsibilities into 2 workmodes:
- pure workmode, which just scans deps and builds (formerly --pure)
- make mode, which does 1) and 2) above, and then writes the Makefile
  with the minified list of dependencies.

the workmode is selected based on the name of the executable.
if it's rcb2make(.py) make mode is selected, else pure mode.
while in theory this removes only a single command line option, in
practice this makes rcb2 now much nicer to work with.

it's recommended to make rcb2make a symlink to rcb2.